Split statement extractor tests into two
